third_party/blink/web_tests/platform/generic/media/W3C/video/networkState/networkState_during_loadstart-expected.txt

videoElement.networkState should be NETWORK_LOADING during loadstart event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "2" is "2"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
